+++ src/main/java/com/example/namo2/domain/memo/MoimMemoLocationRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.example.namo2.domain.memo;
 
-import com.example.namo2.domain.moim.dto.MoimMemoLocationDto;
+import com.example.namo2.domain.moim.ui.dto.MoimMemoLocationDto;
 
 import java.util.List;
