+++ src/main/java/com/example/namo2/domain/memo/MoimMemoLocationRepositoryImpl.java
@@ -3,7 +3,7 @@ package com.example.namo2.domain.memo;
 import com.example.namo2.domain.memo.domain.MoimMemoLocation;
 import com.example.namo2.domain.memo.domain.MoimMemoLocationAndUser;
 import com.example.namo2.domain.memo.domain.MoimMemoLocationImg;
-import com.example.namo2.domain.moim.dto.MoimMemoLocationDto;
+import com.example.namo2.domain.moim.ui.dto.MoimMemoLocationDto;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
 
@@ -16,11 +16,9 @@ import static com.example.namo2.domain.memo.domain.QMoimMemoLocationAndUser.moim
 
 public class MoimMemoLocationRepositoryImpl implements MoimMemoLocationRepositoryCustom {
     private final JPAQueryFactory queryFactory;
-    private final EntityManager em;
 
     public MoimMemoLocationRepositoryImpl(EntityManager em) {
         this.queryFactory = new JPAQueryFactory(em);
-        this.em = em;
     }
 
     @Override
@@ -47,7 +45,7 @@ public class MoimMemoLocationRepositoryImpl implements MoimMemoLocationRepositor
                 .fetch()
                 .stream()
                 .collect(Collectors.groupingBy(moimMemoLocationAndUser -> moimMemoLocationAndUser.getMoimMemoLocation().getId()));
-        moimMemoLocationDtos.stream()
+        moimMemoLocationDtos
                 .forEach((moimMemoLocationDto -> moimMemoLocationDto.addLocationParticipants(locationAndUsersMap.get(moimMemoLocationDto.getMoimMemoLocationId()))));
         return moimMemoLocationDtos;
     }
