+++ src/main/java/com/example/namo2/domain/moim/dao/repository/MoimAndUserRepository.java
@@ -1,4 +1,4 @@
-package com.example.namo2.domain.moim;
+package com.example.namo2.domain.moim.dao.repository;
 
 import com.example.namo2.domain.moim.domain.Moim;
 import com.example.namo2.domain.moim.domain.MoimAndUser;
@@ -13,14 +13,14 @@ import java.util.Optional;
 public interface MoimAndUserRepository extends JpaRepository<MoimAndUser, Long> {
 
     @Query(value = "select gu from MoimAndUser gu join fetch gu.moim where gu.user= :user")
-    public List<MoimAndUser> findMoimAndUserByUser(User user);
+    List<MoimAndUser> findMoimAndUserByUser(User user);
 
     @Query(value = "select gu from MoimAndUser gu join fetch gu.user where gu.moim= :moim")
-    public List<MoimAndUser> findMoimAndUserByMoim(Moim moim);
+    List<MoimAndUser> findMoimAndUserByMoim(Moim moim);
 
-    public Optional<MoimAndUser> findMoimAndUserByUserAndMoim(User user, Moim moim);
+    Optional<MoimAndUser> findMoimAndUserByUserAndMoim(User user, Moim moim);
 
-    public Integer countMoimAndUserByMoim(Moim moim);
+    Integer countMoimAndUserByMoim(Moim moim);
 
-    public boolean existsMoimAndUserByMoimAndUser(Moim moim, User user);
+    boolean existsMoimAndUserByMoimAndUser(Moim moim, User user);
 }
