+++ src/main/java/com/example/namo2/domain/moim/dao/repository/MoimRepository.java
@@ -1,4 +1,4 @@
-package com.example.namo2.domain.moim;
+package com.example.namo2.domain.moim.dao.repository;
 
 import com.example.namo2.domain.moim.domain.Moim;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface MoimRepository extends JpaRepository<Moim, Long> {
-    public Optional<Moim> findMoimByCode(String code);
+    Optional<Moim> findMoimByCode(String code);
 }
