+++ src/main/java/com/example/namo2/domain/moim/dao/repository/MoimScheduleRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.example.namo2.domain.moim;
+package com.example.namo2.domain.moim.dao.repository;
 
 import com.example.namo2.domain.schedule.dto.DiaryDto;
 import com.example.namo2.domain.schedule.dto.SliceDiaryDto;
@@ -8,5 +8,5 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public interface MoimScheduleRepositoryCustom {
-    public SliceDiaryDto<DiaryDto> findMoimScheduleMemoByMonth(Long userId, List<LocalDateTime> dates, Pageable pageable);
+    SliceDiaryDto<DiaryDto> findMoimScheduleMemoByMonth(Long userId, List<LocalDateTime> dates, Pageable pageable);
 }
