+++ src/main/java/com/example/namo2/domain/moim/ui/dto/MoimMemoDto.java
@@ -1,4 +1,4 @@
-package com.example.namo2.domain.moim.dto;
+package com.example.namo2.domain.moim.ui.dto;
 
 import com.example.namo2.domain.memo.domain.MoimMemo;
 import lombok.Getter;
