+++ src/main/java/com/example/namo2/domain/moim/ui/dto/MoimScheduleAlarmDto.java
@@ -1,4 +1,4 @@
-package com.example.namo2.domain.moim.dto;
+package com.example.namo2.domain.moim.ui.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
