+++ src/main/java/com/example/namo2/domain/schedule/ScheduleRepositoryCustom.java
@@ -2,7 +2,7 @@ package com.example.namo2.domain.schedule;
 
 import com.example.namo2.domain.schedule.domain.Schedule;
 import com.example.namo2.domain.user.domain.User;
-import com.example.namo2.domain.moim.dto.MoimScheduleRes;
+import com.example.namo2.domain.moim.ui.dto.MoimScheduleRes;
 import com.example.namo2.domain.schedule.dto.DiaryDto;
 import com.example.namo2.domain.schedule.dto.GetScheduleRes;
 import com.example.namo2.domain.schedule.dto.OnlyDiaryDto;
@@ -23,5 +23,5 @@ public interface ScheduleRepositoryCustom {
 
     List<MoimScheduleRes> findMonthScheduleInMoim(Long moimId, LocalDateTime startDate, LocalDateTime endDate);
 
-    public List<GetScheduleRes> findMoimSchedulesByUserId(User user, LocalDateTime startDate, LocalDateTime endDate);
+    List<GetScheduleRes> findMoimSchedulesByUserId(User user, LocalDateTime startDate, LocalDateTime endDate);
 }
