+++ src/main/java/com/example/namo2/domain/schedule/ScheduleService.java
@@ -8,9 +8,9 @@ import com.example.namo2.domain.moim.domain.MoimSchedule;
 import com.example.namo2.domain.moim.domain.MoimScheduleAndUser;
 import com.example.namo2.domain.schedule.domain.Alarm;
 import com.example.namo2.domain.schedule.domain.Period;
-import com.example.namo2.domain.moim.MoimScheduleAndUserRepository;
-import com.example.namo2.domain.moim.MoimScheduleRepository;
-import com.example.namo2.domain.moim.MoimService;
+import com.example.namo2.domain.moim.dao.repository.MoimScheduleAndUserRepository;
+import com.example.namo2.domain.moim.dao.repository.MoimScheduleRepository;
+import com.example.namo2.domain.moim.application.impl.MoimService;
 import com.example.namo2.domain.schedule.dto.DiaryDto;
 import com.example.namo2.domain.category.domain.Category;
 import com.example.namo2.domain.schedule.domain.Image;
